--- src/main/java/hackathon/microstream/dal/util/CRUDException.java
@@ -1,22 +1,0 @@
-package hackathon.microstream.dal.util;
-
-public class CRUDException extends RuntimeException {
-    public CRUDException() {
-    }
-
-    public CRUDException(String message) {
-        super(message);
-    }
-
-    public CRUDException(String message, Throwable cause) {
-        super(message, cause);
-    }
-
-    public CRUDException(Throwable cause) {
-        super(cause);
-    }
-
-    public CRUDException(String message, Throwable cause, boolean enableSuppression, boolean writableStackTrace) {
-        super(message, cause, enableSuppression, writableStackTrace);
-    }
-}
