+++ src/main/java/hackathon/microstream/dal/util/NotFoundException.java
@@ -1,0 +1,22 @@
+package hackathon.microstream.dal.util;
+
+public class NotFoundException extends RuntimeException {
+    public NotFoundException() {
+    }
+
+    public NotFoundException(String message) {
+        super(message);
+    }
+
+    public NotFoundException(String message, Throwable cause) {
+        super(message, cause);
+    }
+
+    public NotFoundException(Throwable cause) {
+        super(cause);
+    }
+
+    public NotFoundException(String message, Throwable cause, boolean enableSuppression, boolean writableStackTrace) {
+        super(message, cause, enableSuppression, writableStackTrace);
+    }
+}
