+++ src/main/java/hackathon/microstream/service/rest/errorhandling/NotFoundExceptionHandler.java
@@ -1,0 +1,18 @@
+package hackathon.microstream.service.rest.errorhandling;
+
+import hackathon.microstream.dal.util.NotFoundException;
+
+import javax.ws.rs.core.Response;
+import javax.ws.rs.ext.ExceptionMapper;
+import javax.ws.rs.ext.Provider;
+
+@Provider
+public class NotFoundExceptionHandler implements ExceptionMapper<NotFoundException> {
+    @Override
+    public Response toResponse(NotFoundException exception) {
+        return Response
+                .status(Response.Status.NOT_FOUND)
+                .entity("")
+                .build();
+    }
+}
